+++ libs/sabi-core/src/http/header/mod.rs
@@ -1,9 +1,3 @@
-mod header_fragment;
-pub use header_fragment::HeaderFragment;
-
-mod headers;
-pub use headers::Headers;
-
 mod content_type;
 pub use content_type::ContentType;
